backend/requirements-wsl.txt
attrs==22.1.0
beautifulsoup4==4.11.1
branca==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
docopt==0.6.2
Flask==2.2.2
flask-swagger-ui==4.11.1
folium==0.13.0
geopandas==0.11.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pipwin==0.5.2
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pyproj==3.4.0
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
matplotlib
ffmpeg-python
Flask-Cors==3.0.10
xarray